pyproject.toml
[project]
name = "osintsuite"
version = "0.1.0"
description = "OSINT Investigation Suite — CLI + Web platform for open-source intelligence gathering"
requires-python = ">=3.9"
license = "MIT"

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "typer>=0.12",
    "rich>=13.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "python-whois>=0.9",
    "dnspython>=2.6",
    "phonenumbers>=8.13",
    "thefuzz[speedup]>=0.22",
    "weasyprint>=62",
    "python-multipart>=0.0.9",
    "eval_type_backport>=0.2; python_version < '3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "factory-boy>=3.3",
    "ruff>=0.4",
]

[project.scripts]
osint = "osintsuite.cli.app:app"
osint-web = "osintsuite.web.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"
